boot_test.go: split makebootable tests into makebootable_test.go
Smaller test files are easier to maintain and update.
